vim: Improved n and d keymapping
n and d used to be bound to g^ and g$, which turns out to be rather
annoying when lines are only slightly longer than the screen width (as
is usually the case when programming). Thus, these mappings have been
changed to ^ and $, while C-n and C-d are now used for g^ and g$. Their
previous mappings, in turn, are currently not mapped.