fish: better env var handling
a) The PATH is now prepended to only by login shells. This prevents a
minor problem where the PATH would grow with each nested fish
invocation.

b) Most former global variables are now universal since this seems to be
the idiomatic fish way of doing things. Not sure yet if this is actually
an improvement.