shell: Make bash defer to fish in interactive mode
This commit implements the aptly named "bashrc hack" mentioned in the
Gentoo wiki. [1] Essentially, it is not recommended to use fish as the
login shell because it does not respect the usual initialisation files
(/etc/profile, ~/.profile, etc.) and therefore may not set environment
variables correctly. Therefore, we change the default shell to bash and
let it do the environment setup, then `exec fish` at the end of the
.bashrc, which makes fish inherit the environment.

[1] https://wiki.gentoo.org/wiki/Fish#Caveats